refactor: move stories into packages
