+++ .storybook/config.js
@@ -1,4 +1,4 @@
 import { configure } from '@storybook/react';
 
 // automatically import all files ending in *.stories.js
-configure(require.context('../stories', true, /\.stories\.js$/), module);
+configure(require.context('../packages', true, /\.stories\.js$/), module);
