+++ packages/test-button/TestButton.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { action } from '@storybook/addon-actions';
 import { Button } from '@storybook/react/demo';
-import TestButton from '../packages/test-button';
+import TestButton from './index.js';
 
 export default {
   title: 'TestButton',
