+++ packages/test-button/package.json
@@ -2,7 +2,7 @@
   "name": "test-button",
   "version": "0.1.0",
   "private": true,
-  "main": "./lib/index.js",
+  "main": "./dist/index.js",
   "peerDependencies": {
     "react": "^16.12.0",
     "react-dom": "^16.12.0"
